Fix download 860m bug with Pandas 1.3.x
Downloading older 860m files without "archive" in the url leads to an
html file instead of excel file. Pandas used to raise an XLRDError when
trying to read the html (with .xlsx extension). Now it raises a ValueError
or BadZipFile error, depending on if the openpyxl engine is specified.
Fix by adding these error types to the except clause.