Fix find_newest_860m bug
EIA changed their HTML format slightly. Changed the table class in bs4 to fix.
Now raising an error when the table of file names can't be found by scraping.
Also discovered that the lxml library was no longer installed as a
dependency. Added lxml to the environment file to fix.