Big lru_cache push
Use lru_cache for functions that are slow and not too hard to set up properly.

For funcs with "settings" arg, replace with specific kwargs. Do the same in functions they are called from.

Add functions for freezing dataframe inputs. This adds to memory requirements but can save lots of time in some cases.