Handle missing columns in rg data with warning
Missing columns will probably be where there is capacity for a vre resource in a region but no generation profile. This would fail before when trying to access the columns.

Warn the user if columns are missing because it may result in a non-variable generation profile with "1" in all hours.